Use robots.txt for crawler handling
For inventid, we do not want indexing of these images, so we set the variable
`allow_indexing` to `false` (the default). Meanwhile Magnetme might want to
allow indexing for this, so they can override the value in the configuration
json.

this also prevents crawlers from hitting 404s, or longer waiting times.