Properly handle token creation for duplicate tokens
In case a token was already created, `token.js` expected an error to be thrown by `database.createToken`, however that function already handles the exception internally, returning an undefined value to indicate no new token has been created. Because of this `token.js` thought the new token was persisted correctly and returned the unpersisted token. Instead it now return the value as provided by `database.createToken`, so `/token` can indicate a proper 403 here.