Improve restart resiliency
When a database failed, upon the first retry the migration would fail,
but by not correctly passing the error the system would enter a frozen
state. Not serving requests, but not retrying either. By passing the error
properly to the callback, this is fixed.

Additionally when running Docker, this adds a retry policy so it keeps retrying
with an exponential backoff in order to resume service as soon as possible.